requirements.txt

Django==1.9.13
djangorestframework==3.6.2
Pillow==5.4.1
django-filter==1.1.0